% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddat.R
\name{builddat}
\alias{builddat}
\title{Create a dat file}
\usage{
builddat(psi, time = 1:length(psi), depth = 1:nlayers(psi[[1]]),
  overwrite = TRUE, time.inv = time[1:2], elevation, slope, aspect,
  unit_z = 0.001, ...)
}
\arguments{
\item{psi}{... TO DO}
}
\description{
Create a dat file
}

