% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{LateralSubsurfaceDischarge}
\alias{LateralSubsurfaceDischarge}
\alias{LateralSurfaceDischarge}
\title{Lateral Subsurface Discharge}
\usage{
LateralSubsurfaceDischarge(psi, elevation = "DemFile",
  slope = "SlopeMapFile", aspect = NULL, neighbors = 8,
  soilmap = "SoilMapFile", soiltypes = "SoilParFile", Dz = "SoilDz",
  unitelev = 1000, use.soilwater.function = list(fun = "khy", args =
  list(ksat = "LateralHydrConductivity", alpha = "Alpha", n = "N", type_swc =
  "VanGenuchten", type_khy = "Mualem")),
  param.use.soilwater.function = c("LateralHydrConductivity", "Alpha", "N"),
  use_XY = FALSE, unit_input = c("millimeter", "centimeter", "meter"),
  output.discharge = FALSE, ...)

LateralSurfaceDischarge(hsup, elevation = "DemFile", slope = "SlopeMapFile",
  aspect = NULL, neighbors = 8, landmap = "LandCoverMapFile",
  flowres = "SurFlowResLand", expflowres = "SurFlowResExp",
  unit_input = c("millimeter", "centimeter", "meter"),
  output.discharge = FALSE, use_XY = FALSE, ...)
}
\arguments{
\item{psi}{soil presssure head. It is  \code{RasterBrick-class} object.}

\item{elevation, slope, aspect}{elevation  maps of elevation,slope,aspect. Slope and aspect must be expressed in degrees.}

\item{neighbors}{Integer. Indicating how many neighboring cells to use to compute slope for any cell. Either 8 (queen case) or 4 (rook case). Only used for slope and aspect, see \code{\link{terrain}}.}

\item{soilmap}{raster map of soil class types or related keyword of GEOtop file \code{geotop.inpts} (\code{inpts.file} argument in \code{\link{get.geotop.inpts.keyword.value}}.}

\item{soiltypes}{list of soil class types or related keyword of GEOtop file \code{geotop.inpts} (\code{inpts.file} argument in \code{\link{get.geotop.inpts.keyword.value}}. See \code{metacat} arguments in \code{\link{SoilPropertyMap}}.}

\item{Dz}{layer thickness or or related GEOtop keyword used in the \code{soiltypes} file.}

\item{unitelev}{ratio between unis of \code{elevation} and \code{psi}, e.g. if elevetion is exxxpressed in meters and psi in millimeters , it is 1000 (Default).}

\item{use.soilwater.function}{list object passed to \code{\link{SoilPropertyMap}}. It is utilized to calculate unsaturated hydraulic conductivity used in Darcy-Buckingham Law (see also \code{\link{khy}}).}

\item{param.use.soilwater.function}{GEOtop keywords of the soil parameters utilized to calculate unsaturated hydraulic conductivity (see also \code{khy})}

\item{use_XY}{logical vaule. It it \code{TRUE}, the function returns the flux along the X (eastward) and Y (northward) directions.}

\item{...}{further arguments for \code{\link{SoilPropertyMap}}}

\item{hsub}{surfdace water depth. It is  \code{RasterBrick-class} object.}
}
\description{
Lateral Subsurface Discharge

Lateral Surface Discharge - Runoff
}
\examples{
wpath <- system.file("geotop_simulation",package="geotopsim")

#' ## Use soilwater function 
library(soilwater)
data(PsiTheta)

t <- 5
psit <- psi[[t]]
hsupt <- hsub[[t]]

q <- LateralSubsurfaceDischarge(psi=psit,wpath=wpath)
qxy <- LateralSubsurfaceDischarge(psi=psit,wpath=wpath,,output.discharge=TRUE)





qdischarge_sub <- LateralSubsurfaceDischarge(psi=psit,wpath=wpath,output.discharge=TRUE)

qdischarge_sup <- LateralSurfaceDischarge(hsup=hsupt,wpath=wpath,output.discharge=TRUE)

}
\seealso{
\code{\link{terrain}},\code{\link{}},\code{\link{SoilPropertyMap}}
}

