% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpropmap.R
\name{SoilPropertyMap}
\alias{PropertyMap}
\alias{SoilPropertyMap}
\title{Creation of a Physical Variable Map from a Catagory Map}
\usage{
SoilPropertyMap(x, map = "SoilMapFile", metacat = "SoilParFile",
  header = "Header", names = NULL, use.soilwater.function = NULL,
  psi = NULL, ...)

PropertyMap(x, map = "LandUseMapFile", ...)
}
\arguments{
\item{x}{variable or variables to map}

\item{map}{catagory input map, i.e. soiltype maps or related GEOtop keyword.}

\item{metacat}{list of dataframe containingcatagory definitions, e. g. the list of soil properties for each class or related GEOtop keyword.}

\item{header}{header sting using in \code{geotop.inpts} file. Default is \code{"Header"}}

\item{names}{names for the layer of the \code{RasterBrick-class} object returned by the function.}

\item{use.soilwater.function}{list. It it is not \code{NULL}, it contanis a function of \code{soilwater} package which calculated using the variables indicated form \code{x} as input. See examples and details. Default is \code{NULL} nd then ingored.                                                                                                                                  li}

\item{psi}{\code{RasterBrick-class} object containg \code{psi} (pressure head values) use for \code{\link{soilwater}} function in case \code{use.soilwater.function} is not \code{NULL}.}

\item{...}{further arguments for \code{\link{get.geotop.inpts.keyword.value}} such as \code{wpath} or \code{inpts.file}                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       s}
}
\description{
This functions creates physicalvariable maps from catagorical information, e.g. a map of hydraulic conductivity from the map of Soil Categories.
}
\examples{


wpath <- system.file("geotop_simulation",package="geotopsim")
x <- "Alpha"
b <- SoilPropertyMap(x,wpath=wpath)

xp <- c("LateralHydrConductivity","N","Alpha")
bp <- SoilPropertyMap(xp,wpath=wpath)

## Use soilwater function 
library(soilwater)
data(PsiTheta)

t <- 5
psit <- psi[[t]]



#khy(psi = 0.5, v = 0.5, ksat = 0.01, alpha = 1, n = 1.5, m = 1 -
#  1/n, theta_sat = 0.4, theta_res = 0.05, psi_s = -1/alpha, lambda = m *
#  n, b = NA, type_swc = "VanGenuchten", type_khy = c("Mualem",
#  "BrooksAndCorey"), ...)


use.soilwater.function <- list()
use.soilwater.function$fun <- "khy" ## called function of "soilwater" packege
use.soilwater.function$args <- list(ksat="LateralHydrConductivity",alpha="Alpha",n="N", 
								  type_swc = "VanGenuchten", type_khy = "Mualem")


xs <- c("LateralHydrConductivity","Alpha","N")

khmap <- SoilPropertyMap(xs,wpath=wpath,use.soilwater.function=use.soilwater.function,psi=psit)



### BELOW a list of some keywords used for soil property variables
#HeaderSoilDz	=	"Dz"
#HeaderLateralHydrConductivity	=	"Kh"
#HeaderNormalHydrConductivity	=	"Kv"
#HeaderThetaRes	=	"vwc_r"
#HeaderWiltingPoint	=	"vwc_w"
#HeaderFieldCapacity	=	"vwc_fc"
#HeaderThetaSat	=	"vwc_s"
#HeaderAlpha	=	"alpha"
#HeaderN	=	"n"
#HeaderSpecificStorativity	=	"stor"
#HeaderKthSoilSolids	=	"Kth"
#HeaderCthSoilSolids	=	"Cth"
#HeaderSoilInitPres = "InitPsi"
}

