% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilWaterStorage.R
\name{SoilWaterStorage}
\alias{SoilWaterStorage}
\title{Soil Moisture Mapping}
\usage{
SoilWaterStorage(theta, psi, layers = 1, fun = sum, na.rm = TRUE,
  psi_thres = 0, comparison = c("<", ">", "<=", ">=", "==", "!="),
  indices = 1, bedrock.depth = NULL, ...)
}
\arguments{
\item{theta}{soil water content \code{RasterBrick-class} object}

\item{psi}{soil pressure head content \code{RasterBrick-class} object}

\item{layers}{layer thickness. It is a scalar or a vector whose length is equal to the number of \code{theta} layers.}

\item{fun}{function aggregation}

\item{na.rm}{a logical value for \code{fun} indicating whether NA values should be stripped before the computation proceeds.}

\item{psi_thres}{threshold value.  It is 0 for separating satureted and unsatureted zones, or \code{NA} if the two zones are not distinguished.}

\item{comparison}{comparison operator for \code{psi} .See \code{\link{Comparison}}. Default is the first element of \code{c("<",">","<=",">=","==","!=")}}

\item{indices}{see \code{\link{stackApply}}}

\item{bedrock.depth}{map of deprock depth. If it is not \code{NULL}, \code{theta} cells under bedrock are not considered.}

\item{...}{further aguments for \code{\link{stackApply}}}
}
\description{
Vertical or Slope-Normal Integration/Aggregation of soil water content
}
\examples{

data(PsiTheta)

 t <- 5
 dz <- 50 ## 50 millimiters layer depth
 unsatWaterVolume <- SoilWaterStorage(theta[[t]],psi[[t]],layer=dz,comparison="<",psi_thres=0)
 satWaterVolume <- SoilWaterStorage(theta[[t]],psi[[t]],layer=dz,comparison=">=",psi_thres=0)







}

